requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utils
tqdm>=4.66.0
matplotlib>=3.7.0
streamlit>=1.38.0
